+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* updated GitLab CE/Enterprise to use GitLab API v4
 
 #### 8.1.0 / 2018-06-26
 * added `register_activation_hook` to correctly rename directory to `github-updater` on activation; activation will fail if rename successful.